Update true color detection
Since this plugin was created Vim has released the 'termguicolors'
option, after it's inclusion Neovim deprecated
'NVIM_TUI_ENABLE_TRUE_COLOR' in favor of 'termguicolors'.

This removes the deprecated Neovim options and adds a check for
'termguicolors'. The only side effect is it will no longer work on
outdated versions of Neovim.