Replace custom highlights with a default highlight
This removes all the custom highlight setup functions and replaces it
with a default highlight command. This allows configuration in the form
of a highlight command instead of setting a variable. It will break
backwards compatibility with the g:qs_first_occurrence_highlight_color
and g:qs_second_occurrence_highlight_color options.

I feel this is a simpler more robust solution to creating highlights and
that it will avoid or solve issues related to varying levels of color
support.