Remove final calls to old highlight functions
This removes the last calls to add_to_highlight_group, but this is a
tricky case. The plugin switched the secondary highlight group to the
colors of the primary group, it then switched them back. Now that the
plugin is using highlight groups directly this becomes a lot harder to
do while maintaining customization.

I have started a solution to this by attempting to save the setting of
QuickScopeSecondary before and reset to that after.