Add verbose check for saving highlight group
If verbose is set the highlight command will also print where it was
last set from which would mess up the reset command. This saves the
value of verbose before and restores it after saving the value of the
highlight group.

This also adds silent to the command to save the highlight group which
avoids the "Press <Enter>..." message making the redraw unnecessary.

Finally it corrects a typo where the secondary highlight group was being
linked to itself in the double_tap function. However, this link to the
primary group still is not visible like it is in the current master
branch.