Fix mappings that clobber select-mode
Fixes: https://github.com/unblevable/quick-scope/issues/25

For the main internal mappings of the plugin `noremap` was used, which
by default maps in Normal, Visual, Select and Operator-pending modes
(see `:help map-modes`). These mappings only seem to be needed in Normal
mode, so this was changed to a `nnoremap`.

Also the `<plug>(QuickScopeToggle)` function was mapped with `vnoremap`
which maps in Visual and Select modes. This was changed to a `xnoremap`.