Re-enable plugin in visual and operator-pending
In 85afb9b04178941d377c7fcd6295e7e6704940b0 the plugin was restricted to
normal mode in order to fix select mode. Having the plugin work in
operator pending mode is useful for commands like `cf<char>` or
`dt<char>`, and visual mode may also be useful.

This fixes it by adding a for loop over the modes we want the mappings
to work in.