Fix clobbering users '<c-c>' mappings
Resolve https://github.com/bradford-smith94/quick-scope/issues/2

Previously the plugin clobbered users '<c-c>' mappings in order to exit
quietly in the middle of a quick-scope movement. This utilizes a
simplified version of the solution posed in [1] to save and restore any
mappings.

The new function is added to 'autoload/quick_scope.vim' in order to
prepare for a later update moving the bulk of the functions to autoload.

[1]: https://vi.stackexchange.com/a/7735