Fix Vim freeze when using Vanilla-mode
Resolve https://github.com/bradford-smith94/quick-scope/issues/1

When using quick-scope in the Vanilla-mode, which always highlights
lines, v2.3.0 caused a freeze. Vim could be unfrozen by using '<c-c>'
but would freeze again upon moving up or down a line. The same freeze
could be observed using the `highlight_on_keys` option if one started a
backwards search (F/T) at the beginning of a line.

This resolves the issue by removing a line which decremented the cursor
position prior to sending it to the main highlighting function
'get_highlight_patterns'. I'm not completely sure why the decrement was
there to begin with but it seems the way that 'get_highlight_patterns'
was done in the multibyte merge that it doesn't expect the decrement to
be there.

I don't believe that removing this will cause any other issues, and in
my brief testing right now it doesn't appear to.