Move functions to autoload
There are a lot of functions for this plugin and if one is using the
'highlight on key press' mode they are not needed when Vim is starting
up. Moving these functions to the autoload directory will save startup
time. However, if using the 'vanilla' mode these functions will probably
end up being loaded fairly quickly anyway.

I chose to keep the mapping restore function in a separate autoload file
thinking that I would try to split the functions further based on what
is needed by the two highlighting modes. Since these modes share most of
the core highlighting functions it is harder to split them up than I
thought it would be.