Move targets check to detect words better
Fixes: https://github.com/unblevable/quick-scope/issues/10

This issue was that quick-scope did not properly detect word
separations. At first I thought this was because it was checking for
space characters but the change in
d40efc96cc7e3f1a3fadbbc8b22c2e37c46ff81a didn't completely fix the
issue.

This is because of this call to index(). Originally quick-scope was
marking words whenever it found something that wasn't a target character
(by default targets are a-z,A-z and 0-9). So any other character (no
matter where in a word it was) made quick-scope think it reached a new
"word".

This changes that by moving and negating the call to the else case. So
now quick-scope will detect the start of a new word whenever it reaches
a character that is not in 'iskeyword' (see `:help 'iskeyword'`) or an
empty character, for any other character quick-scope only considers it
for highlighting if it is present in the list of target characters.

There is no longer an else case for this, any non-target (but still
keyword i.e. part of a word) character is simply skipped by quick-scope.