Fix E714 in quick_scope#DoubleTap()
Resolves: https://github.com/bradford-smith94/quick-scope/issues/3

When `quick_scope#DoubleTap()` ends up being called an E714 (List
required) ends up being triggered, which completely breaks the 'f/F/t/T'
motion being done.

In `quick_scope#DoubleTap()` it was found that
`quick_scope#HighlightLine()` was being called with 'targets' as a
Dictionary. This was the old style, in the multibyte merge
(6fc165632b98be37d60ec8ba83220e761436f298) 'targets' (along with
g:qs_accepted_chars) was changed to a List as the value stored in the
Dictionary was never used. In the merge the function call in
`quick_scope#DoubleTap()` was overlooked.

This fixes the E714 by changing the argument from constructing a
Dictionary with a key of the target and value of empty string, to
constructing a list with the element of the target.