Add backwards compatibility for highlight config
This adds a backwards compatibility check for the old style highlighting
options. This should add a default highlight for both the highlight
groups with underline and the specified color. Since it is a default
highlight it can still be overridden by another highlight command.

This will also print an error message telling the user that the old
style highlight options are now deprecated. Currently this is super
annoying, as it prints while vim is starting so you see it on the
command line before Vim fully starts. I guess this is fine, the option
is deprecated anyway, but perhaps there is a better place to move the
echo line to.