Add detail to custom colors documentation
This adds more detail to the documentation in the README and the Vim
help page about setting up custom highlight colors. This is because if
the autocmd group is set after setting colorscheme in the .vimrc then
the highlight groups will not work properly (as the ColorScheme autocmd
event will have already triggered).

Adding these lines should (hopefully) draw more attention to putting the
autocmd group **before** setting colorscheme.