Change echoerr to echomsg for deprecation warnings
The deprecated setting error messages are not really errors so much as
they are warnings to get the user's attention. As such we don't need Vim
to print an error with a line number. This replaces the instances of
`echoerr` with `echomsg` and makes use of `echohl` to color the message
like an error message would be colored. The result of this is a
(slightly) less intrusive error message.