Add more autocmd events to vanilla mode triggers
https://github.com/unblevable/quick-scope/issues/42

This was inspired by @linusboyle's fix for the issue. These extra
autocmd events should handle a few more cases where the issue may come
up again and with the addition of BufEnter on the re-highlight trigger
we get the added benefit of a quicker highlight response upon entering a
buffer.

This also adds to the help page changelog.