Fix issue restoring highlight
Resolves https://github.com/unblevable/quick-scope/issues/46

The substitute call inside s:reset_saved_secondary_highlight() changed
the value of the s:saved_secondary_highlight variable, this caused the
reset function to fail if it was called multiple times before the save
function was called again.

This fix changes the variable name so that it doesn't overwrite the
variable that controls the flow of the reset function.