Add option to slow down vanilla highlight
Resolves https://github.com/unblevable/quick-scope/issues/8

This resolves the slowdown issue by adding an option to use the
CursorHold event as opposed to the CursorMoved event. For those who
don't experience the slowdown, nothing has changed, in the case you are
affected by the slowdown you can now set `g:qs_lazy_highlight = 1` in
order to use the CursorHold event instead.