Only apply v:count1 when highlight_on_keys is set
Fix https://github.com/unblevable/quick-scope/issues/60

With the Vanilla always on highlight mode we have no way of knowing at
highlight time whether the user will type `[count]f*` or just `f*` so we
have to highlight assuming count of 1.

This adds a conditional to check for the highlight_on_keys variable to
enable using count for highlighting the appropriate [count] targets.