Make check for disables more clear
Fixes https://github.com/unblevable/quick-scope/issues/75

The mentioned issue points out a missing "< 0" in the conditional that
checks if highlighting should be disabled. This if statement had grown
quite long and as a result it was easier to miss this kind of a mistake.

The issue is resolved by wrapping the line so that it is more clear all
the checks that are being made (in addition to the missing "< 0" being
added).

The current checks intended are as follows:
- the g:qs_enable variable is true (enabling Quickscope)
- AND the b:qs_local_disable either hasn't been set or is false
- AND the current buftype is not in g:qs_buftype_blacklist
- AND the current filetype is not in g:qs_filetype_blacklist