Make mapping reset handle multiple modes
Fix https://github.com/unblevable/quick-scope/issues/87

In some cases `maparg()` will return a string of multiple characters for
multiple mapping modes, one case will be a `noremap` followed by a
different `vnoremap` resulting in a map mode of 'no'. The previous
solution failed in this case because 'nonoremap' is not a proper
command.

This fix creates a loop to go through that string character by character
in order to restore the mapping. It also adds handling for special cases
(' ' for 'nvo' and '!' for 'ic' modes).