Make operating mode checking more fixed between plugin and autoload
https://github.com/unblevable/quick-scope/issues/98

The plugin currently configures it's operating mode when the plugin
script file is sourced. After this there is no support for changing
operating mode.

Previously the autoload functions had two locations where they checked
the value of `g:qs_highlight_on_keys` to determine which operating mode
the plugin was running in, this introduced a "time of set to time of
check" issue where the config could set the value after the plugin has
been initialized. This is an issue because some functions used only by
highlight on keys mode initialize variables used later in the
highlighting process, which are uninitialized if the plugin is set up in
"vanilla" mode but the highlighting process later thinks it's in "keys"
mode. Which is the issue seen in issue 98.

This change resolves the errors by adding a function to set a script
local variable with the operating mode. This value is set when the
plugin is initialized and never changed, so that the highlighting
process does not have to worry about operating mode changing.